Commit final

+++ src/main/java/application/record/LivroDTO.java
@@ -2,7 +2,7 @@ package application.record;
 
 import application.model.Livro;
 
-public record LivroDTO(int id, String titulo, String generos, String autores) {
+public record LivroDTO(Long id, String titulo, String generos, String autores) {
     public LivroDTO(Livro livro) {
         this(livro.getId(), livro.getTitulo(), livro.getGeneros(), livro.getAutores());
     }
